Add core PatchMode functionality
This adds a PatchMode to zextractor. In PatchMode
a filetable is being created that contains the list
of files that are being extracted together with the
name of the origin file (GRF, GPF, THOR, custom name).
This filetable is then read when extracting more files
of different GRF/GPF/THOR files which are supposed to be
patch files. When extracting patch files a priority of
origin files are required which is being used to determine
if the contents of the patch files are supposed to be extracted
or not.

For example if we do an initial extraction of all files in
rdata.grf and data.grf, and we wish to apply patches from
standard RO then we do not wish to replace files which are
from the rdata.grf (renewal). As this would break the priority
as defined in the DATA.ini file.